+++ e2e/flows/add_product_to_cart.js
@@ -6,13 +6,13 @@ export default async function(page) {
   await page.waitForTimeout(2000);
 
   // choose first item
-  await page.locator('main#main ul li a').first().click();
+  await page.locator('#main .products li').first().click();
   await page.waitForTimeout(2000);
 
   // add to cart, take to checkout
   await page.getByRole('button', { name: 'Add to cart' }).click();
   await page.waitForTimeout(2000);
-  await page.locator('#content').getByRole('link', { name: 'View cart' }).click();
+  await page.locator('.woocommerce-message').getByRole('link', { name: 'View cart' }).click();
   await page.getByRole('link', { name: 'Proceed to Checkout' }).click();
   await page.waitForTimeout(2000);
 }