+++ e2e/tests/mlb/chocustom/credit_card_payment_binary_on.spec.js
@@ -1,0 +1,14 @@
+import { test, expect } from "@playwright/test";
+import { mlb } from "../../../data/meli_sites";
+import fillStepsToCheckout from "../../../flows/fill_steps_to_checkout";
+import payWithCard from "../../../flows/mlb/pay_with_card";
+
+const{ url, credit_card_scenarios, guestUserMLB } = mlb;
+const { PENDING } = credit_card_scenarios;
+
+test('test pending payment as guest with master, payment must be rejected and decline message must be shown', async ({page}) => {
+  await fillStepsToCheckout(page, url, guestUserMLB);
+  await payWithCard(page, PENDING.master, PENDING.form);
+
+  await expect(page.locator('div.wc-block-components-notices .wc-block-store-notice')).toHaveText(/The card issuing bank declined the payment/i);
+});