+++ src/WoocommerceMercadoPago.php
@@ -33,7 +33,7 @@ if (!defined('ABSPATH')) {
 
 class WoocommerceMercadoPago
 {
-    private const PLUGIN_VERSION = '7.9.4';
+    private const PLUGIN_VERSION = '7.10.0';
 
     private const PLUGIN_MIN_PHP = '7.4';